fastapi/requirements.txt
absl-py
aiofiles
annotated-types
anyio
astunparse
audioread
certifi
cffi
charset-normalizer
click
decorator
dnspython
email_validator
fastapi
fastapi-cli
flatbuffers
gast
google-pasta
greenlet
grpcio
h11
h5py
httpcore
httptools
httpx
idna
joblib
keras
lazy_loader
librosa
llvmlite
Markdown-it-py
MarkupSafe
mdurl
ml-dtypes
msgpack
namex
noisereduce
numba
numpy
opt-einsum
optree
orjson
packaging
pooch
protobuf
psycopg2-binary
pycparser
pydantic_core
Pygments
python-dotenv
python-multipart
PyYAML
requests
rich
scikit-learn
scipy
shellingham
six
sniffio
soundfile
soxr
SQLAlchemy
starlette
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
typer
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
wrapt
pyjwt
pydub
bcrypt